Landing page examples
Both are responsive and good at both desktop and mobile devices viewing.

First one is a single landing page with simple texts, video, and it's like an about me page.
Second one is a three-page site, first is homepage, second shows what the content is about, and third is a signup form. All pages are really responsive and good to view on devices with 330px and up.